backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.9.2
